Fix a few bugs
- Fix a bug caused by the optimization of CachedRecord
- Add a check in CalvinCacheMgr to avoid updates by read-only transactions

+++ src/main/java/org/elasql/cache/calvin/CalvinCacheMgr.java
@@ -21,6 +21,7 @@ import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.LinkedBlockingQueue;
 
 import org.elasql.cache.CachedRecord;
+import org.elasql.cache.CachedRecordBuilder;
 import org.elasql.cache.VanillaCoreCrud;
 import org.elasql.server.Elasql;
 import org.elasql.sql.RecordKey;
@@ -121,11 +122,17 @@ public class CalvinCacheMgr {
 	}
 
 	public void update(RecordKey key, CachedRecord rec) {
+		if (tx.isReadOnly())
+			throw new RuntimeException("Tx." + tx.getTransactionNumber() + " is read-only.");
+		
 		rec.setSrcTxNum(tx.getTransactionNumber());
 		cachedRecords.put(key, rec);
 	}
 
 	public void insert(RecordKey key, Map<String, Constant> fldVals) {
+		if (tx.isReadOnly())
+			throw new RuntimeException("Tx." + tx.getTransactionNumber() + " is read-only.");
+		
 		CachedRecord rec = new CachedRecord(key, fldVals);
 		rec.setSrcTxNum(tx.getTransactionNumber());
 		rec.setNewInserted(true);
@@ -133,13 +140,19 @@ public class CalvinCacheMgr {
 	}
 
 	public void insert(RecordKey key, CachedRecord rec) {
+		if (tx.isReadOnly())
+			throw new RuntimeException("Tx." + tx.getTransactionNumber() + " is read-only.");
+		
 		rec.setSrcTxNum(tx.getTransactionNumber());
 		rec.setNewInserted(true);
 		cachedRecords.put(key, rec);
 	}
 
 	public void delete(RecordKey key) {
-		CachedRecord dummyRec = new CachedRecord(key);
+		if (tx.isReadOnly())
+			throw new RuntimeException("Tx." + tx.getTransactionNumber() + " is read-only.");
+		
+		CachedRecord dummyRec = new CachedRecordBuilder(key).build();
 		dummyRec.setSrcTxNum(tx.getTransactionNumber());
 		dummyRec.delete();
 		cachedRecords.put(key, dummyRec);
